Define addition, subtraction, and weighting operations for vectors
This adds methods to the base model class for adding, subtracting, and computing the weighted average of state/measurement vectors so that any normalization/constraint operations can be applied here. All the filters then simply call these methods for all vector arithmetic, and all normalizations will be taken care of.