Monitor if the system is stationary
Enables the system/measurement models to determine if the state is
stationary, and update accordingly (e.g. for performing a ZUPT).
The filter will skip a prediciton/correction when stationary.

This adds to the base model class methods for checking if the state is
stationary and modifying it to be stationary (e.g. zero velocities).
This functionality can be turned on/off in each model.