Enable only updating a subset of the state
This modifies all the models and algorithms to only update a subset of the state, and operate on a subset of the controls or measurements.

The motivation for this is to enable defining a generalized state representation and update equations that can be applied to multiple applications. For example, for a 6DOF system the state may need to contain acceleration depending on if the accelerometer data is used as a control input or a measurement. Instead of defining the same measurement model for multiple state representations, a single one can be used (e.g. one odometry measurement model for the state [x, y, Vx, Vy], and another model for the state [x, y, Vx, Vy, Ax, Ay]).

The changes are as follows:
-Each model maintains a list of active states and controls/measurements
-Derived models should only update the active states and use the active controls/measurements
-The filter algorithms operate only on the subset of the state data
-Any state elements that are not "active" will not be altered during updates