Create variable and fixed state sizes of each filter
There are now two version of each fiter: one that operates on the full 
dimensionality of the state, control, and measurement vectors, and 
another that has logic for only operating on subsets of the these 
vectors. The latter versions have 'VS' appended to their names.

Operating on only a subset of the vectors can have a performance impact 
from all the copying operations. This will add the option to avoid such 
overhead if the full dimensionality will be used.