Retain control covariance in filter inputs/states
Since control covariance was previously not retained in the FilterInput 
struct, if that covariance changed between inputs it would not be 
properly accounted for when queuing controls or rewinding the state.

Added the control covariance as an extra element, and updated all the 
predict/correct methods to always query the control covariance from the 
system model and store it.